test: try to use different classloaders to load jdbc drivers
