+++ src/main/java/com/google/cloud/spanner/pgadapter/statements/JdbcMetadataStatementHelper.java
@@ -24,7 +24,7 @@ import com.google.cloud.spanner.pgadapter.utils.PgJdbcCatalog;
  * <p>This class detects whether a statement is generated by a JDBC metadata method and replaces it
  * with a compatible query.
  */
-public class JdbcMetadataStatementHelper {
+class JdbcMetadataStatementHelper {
 
   /**
    * Indicates whether a SQL statement could potentially be a JDBC metadata query that needs
@@ -38,52 +38,16 @@ public class JdbcMetadataStatementHelper {
     if (!sql.contains("pg_catalog.")) {
       return false;
     }
-    // TODO: Register all known JDBC metadata query prefixes from all 42.x versions of the JDBC
-    // driver and only return true if this query starts with one of those.
     return true;
   }
 
-  static String replaceJdbcMetadtaStatement(String sql) {
+  static String replaceJdbcMetadataStatement(String sql) {
     if (sql.startsWith(PgJdbcCatalog.PG_JDBC_EXPORTED_IMPORTED_KEYS_PREFIX)
         || sql.startsWith(PgJdbcCatalog.PG_JDBC_EXPORTED_IMPORTED_KEYS_42_0_PREFIX)) {
-      String replacedSql = PgJdbcCatalog.PG_JDBC_EXPORTED_IMPORTED_KEYS_REPLACEMENT;
-      int startIndex;
-      if (sql.contains(" AND pkn.nspname = ")) {
-        startIndex = sql.indexOf(" AND pkn.nspname = ");
-      } else if (sql.contains(" AND fkn.nspname = ")) {
-        startIndex = sql.indexOf(" AND fkn.nspname = ");
-      } else if (sql.contains(" AND pkc.relname = ")) {
-        startIndex = sql.indexOf(" AND pkc.relname = ");
-      } else if (sql.contains(" AND fkc.relname = ")) {
-        startIndex = sql.indexOf(" AND fkc.relname = ");
-      } else {
-        startIndex = sql.indexOf(" ORDER BY ");
-      }
-      replacedSql += sql.substring(startIndex);
-      return replacedSql
-          .replace(" AND pkn.nspname = ", " AND PARENT.TABLE_SCHEMA = ")
-          .replace(" AND fkn.nspname = ", " AND CHILD.TABLE_SCHEMA = ")
-          .replace(" AND pkc.relname = ", " AND PARENT.TABLE_NAME = ")
-          .replace(" AND fkc.relname = ", " AND CHILD.TABLE_NAME = ")
-          .replace(
-              " ORDER BY fkn.nspname,fkc.relname,con.conname,pos.n",
-              " ORDER BY CHILD.TABLE_CATALOG, CHILD.TABLE_SCHEMA, CHILD.TABLE_NAME, KEY_SEQ")
-          .replace(
-              " ORDER BY pkn.nspname,pkc.relname, con.conname,pos.n",
-              " ORDER BY PARENT.TABLE_CATALOG, PARENT.TABLE_SCHEMA, PARENT.TABLE_NAME, KEY_SEQ");
+      return replaceImportedExportedKeysQuery(sql);
     }
     if (sql.startsWith(PgJdbcCatalog.PG_JDBC_GET_SCHEMAS_PREFIX)) {
-      String replacedSql = PgJdbcCatalog.PG_JDBC_GET_SCHEMAS_REPLACEMENT;
-      int startIndex;
-      if (sql.contains(" AND nspname LIKE ")) {
-        startIndex = sql.indexOf(" AND nspname LIKE ");
-      } else {
-        startIndex = sql.indexOf(" ORDER BY ");
-      }
-      replacedSql += sql.substring(startIndex);
-      return replacedSql
-          .replace(" AND nspname LIKE ", " AND schema_name LIKE ")
-          .replace(" ORDER BY TABLE_SCHEM", " ORDER BY schema_name");
+      return replaceGetSchemasQuery(sql);
     }
     if (sql.startsWith(PgJdbcCatalog.PG_JDBC_GET_FUNCTIONS_WITH_FUNC_TYPE_PREFIX)
         || sql.startsWith(PgJdbcCatalog.PG_JDBC_GET_FUNCTIONS_WITHOUT_FUNC_TYPE_PREFIX)) {
@@ -119,21 +83,10 @@ public class JdbcMetadataStatementHelper {
 
     // Rewrite the 42.3.x PK query to the version in 42.2.x.
     if (sql.startsWith(PgJdbcCatalog.PG_JDBC_PK_QUERY_PREFIX_42_3)) {
-      String replacedSql = PgJdbcCatalog.PG_JDBC_PK_QUERY_PREFIX;
-      String whereClause = "";
-      if (sql.contains(" AND n.nspname = ")) {
-        int startIndex = sql.indexOf(" AND n.nspname = ");
-        int endIndex = sql.indexOf(" AND ct.relname = ", startIndex);
-        whereClause += sql.substring(startIndex, endIndex);
-      }
-      if (sql.contains(" AND ct.relname = ")) {
-        int startIndex = sql.indexOf(" AND ct.relname = ");
-        int endIndex = sql.indexOf(" AND i.indisprimary ", startIndex);
-        whereClause += sql.substring(startIndex, endIndex);
-      }
-      sql = replacedSql + whereClause + " ORDER BY ct.relname, pk_name, key_seq";
+      sql = replacePG42_3_PkQuery_With_42_2_Query(sql);
     }
 
+    // Replace fixed query prefixes and known unsupported tables.
     return sql
         // Replace fixed query prefixes.
         .replace(PgJdbcCatalog.PG_JDBC_PK_QUERY_PREFIX, PgJdbcCatalog.PG_JDBC_PK_QUERY_REPLACEMENT)
@@ -181,5 +134,64 @@ public class JdbcMetadataStatementHelper {
 
         // Replace unsupported casts.
         .replaceAll("'pg_class'::regclass", "0");
+  }
+
+  /** Replaces the getPrimaryKeys query in PG JDBC 42.3.x with the one in 42.2.x, so we can base the further query replacement on the 42.2.x version. */
+  private static String replacePG42_3_PkQuery_With_42_2_Query(String sql) {
+    String replacedSql = PgJdbcCatalog.PG_JDBC_PK_QUERY_PREFIX;
+    String whereClause = "";
+    if (sql.contains(" AND n.nspname = ")) {
+      int startIndex = sql.indexOf(" AND n.nspname = ");
+      int endIndex = sql.indexOf(" AND ct.relname = ", startIndex);
+      whereClause += sql.substring(startIndex, endIndex);
+    }
+    if (sql.contains(" AND ct.relname = ")) {
+      int startIndex = sql.indexOf(" AND ct.relname = ");
+      int endIndex = sql.indexOf(" AND i.indisprimary ", startIndex);
+      whereClause += sql.substring(startIndex, endIndex);
+    }
+    return replacedSql + whereClause + " ORDER BY ct.relname, pk_name, key_seq";
+  }
+
+  private static String replaceGetSchemasQuery(String sql) {
+    String replacedSql = PgJdbcCatalog.PG_JDBC_GET_SCHEMAS_REPLACEMENT;
+    int startIndex;
+    if (sql.contains(" AND nspname LIKE ")) {
+      startIndex = sql.indexOf(" AND nspname LIKE ");
+    } else {
+      startIndex = sql.indexOf(" ORDER BY ");
+    }
+    replacedSql += sql.substring(startIndex);
+    return replacedSql
+        .replace(" AND nspname LIKE ", " AND schema_name LIKE ")
+        .replace(" ORDER BY TABLE_SCHEM", " ORDER BY schema_name");
+  }
+
+  private static String replaceImportedExportedKeysQuery(String sql) {
+    String replacedSql = PgJdbcCatalog.PG_JDBC_EXPORTED_IMPORTED_KEYS_REPLACEMENT;
+    int startIndex;
+    if (sql.contains(" AND pkn.nspname = ")) {
+      startIndex = sql.indexOf(" AND pkn.nspname = ");
+    } else if (sql.contains(" AND fkn.nspname = ")) {
+      startIndex = sql.indexOf(" AND fkn.nspname = ");
+    } else if (sql.contains(" AND pkc.relname = ")) {
+      startIndex = sql.indexOf(" AND pkc.relname = ");
+    } else if (sql.contains(" AND fkc.relname = ")) {
+      startIndex = sql.indexOf(" AND fkc.relname = ");
+    } else {
+      startIndex = sql.indexOf(" ORDER BY ");
+    }
+    replacedSql += sql.substring(startIndex);
+    return replacedSql
+        .replace(" AND pkn.nspname = ", " AND PARENT.TABLE_SCHEMA = ")
+        .replace(" AND fkn.nspname = ", " AND CHILD.TABLE_SCHEMA = ")
+        .replace(" AND pkc.relname = ", " AND PARENT.TABLE_NAME = ")
+        .replace(" AND fkc.relname = ", " AND CHILD.TABLE_NAME = ")
+        .replace(
+            " ORDER BY fkn.nspname,fkc.relname,con.conname,pos.n",
+            " ORDER BY CHILD.TABLE_CATALOG, CHILD.TABLE_SCHEMA, CHILD.TABLE_NAME, KEY_SEQ")
+        .replace(
+            " ORDER BY pkn.nspname,pkc.relname, con.conname,pos.n",
+            " ORDER BY PARENT.TABLE_CATALOG, PARENT.TABLE_SCHEMA, PARENT.TABLE_NAME, KEY_SEQ");
   }
 }
